pyproject.toml
[tool.poetry]
name = "tgen"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^9.3.0"
opensimplex = "^0.4.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
